refactor: Move priorityQueue.ts and its tests
Create src/dataStructures. Move priorityQueue.ts there. Split its tests
into unit tests and property-based tests and move them into the
directory tests/.
